+++ PowerPointLabs/PowerPointLabs/ActionFramework/Action/RemoveHighlightingActionHandler.cs
@@ -2,6 +2,7 @@
 using PowerPointLabs.ActionFramework.Common.Attribute;
 using PowerPointLabs.ActionFramework.Common.Extension;
 using PowerPointLabs.ActionFramework.Common.Interface;
+using PowerPointLabs.HighlightLab;
 
 namespace PowerPointLabs.ActionFramework.Action
 {
