+++ PowerPointLabs/PowerPointLabs/HighlightLab/HighlightTextFragments.cs
